python_basic/requirements.txt
Bottleneck==1.3.5
certifi==2022.6.15
dnspython==1.16.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.3
numpy==1.22.3
packaging==21.3
pandas==1.4.3
pip==22.1.2
pymongo==4.2.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
setuptools==61.2.0
six==1.16.0
wheel==0.37.1
